Block SIGINT during bootstrap
This change is to catch SIGINT during bootstrap
for all providers. Also, if SIGINT is delivered
while tools are being uploaded, the upload will
be cancelled.

If the user Ctrl-C's while the local provider's
bootstrap script is being executed, it will be
interrupted and will return and error to juju.
Juju will then attempt to destroy the environment
and remove the .jenv file.

We will now also send some very basic feedback to
stderr when uploading tools.

Fixes lp:1296475