rate limit non-user login requests (bug #1290843)
this changes the Login request, so that if there are more than N concurrent requests,
it bounces the Login request, rather than letting them pile up.
Even when Machines and Units are rate limited, we let User's Login. This should
keep responsiveness reasonable, even when we have 1000s of units and they all
get bounced.