Introduce instance.Placement
This change introduces a new Placement
structure in the instance package that
encapsulates machine-id, container, and
(new) environment-specific placement
directives.

The add-machine subcommand is modified
to pass this new Placement structure
through the AddMachine client API, which
calls a new state method AddMachineWithPlacement.

Machine-id and container placement do the
same thing as before, while new environment-
specific placement structures are stored in
state. A new state policy is added to allow
an Environ to validate placement.

The provisioner API and worker are updated
to pass Placement through to StartInstance.

Follow-ups to come:
 - update MAAS provider to pick maas-name out
   the Placement.
 - update ec2, OpenStack (and MAAS?) to pick
   out availbility-zone from Placement.
 - update add-unit and deploy subcommands to
   use Placement; this will involve modifications
   to state unit assignment to choose existing
   matching machines (e.g. for maas-name).