state: promote/demote available/unavailable state servers
When EnsureAvailability is called, we now check the
availability of existing state servers. If a state server
is unavailable, then we either set WantsVote=false, or
remove it entirely if it doesn't want a vote.

If we need to create new state servers, we will first
promote existing, available, non-voting state servers
before creating new ones.