refactor apiRootForEntity
It was set up as a way to have custom roots, but we weren't actually
using that, and we were having some really odd side effects of calling
it. (Setting up pingers, etc, which is fine for Login, but really
weird for apiRootForEntity.)

So for now we just get rid of it, since we need the concrete
type in Login.