worker/instancepoller: TestBatching less race-y
The existing TestBatching test was very prone to
failure due to the nature of the test. The test
ensures that requests to Instances are batched
using a token bucket.

Whereas the existing test streamed requests, the
new implementation compresses the addition of
requests into the point where we service an
Instances call.

Fixes lp:1319652